% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOBO.R
\name{detect_hoboware_configuration}
\alias{detect_hoboware_configuration}
\title{Detect HOBOWare file properties}
\usage{
detect_hoboware_configuration(filename)
}
\arguments{
\item{filename}{character, path to the file}
}
\value{
named list of HOBOWare configuration properties (see details for names)
}
\description{
Auto-detect the various HOBOWare text file configuration options that were set for a file
}
\details{
The name keywords to configure a HOBOWare file are:

\itemize{
\item \code{include_line_number}: whether the line number is included as a column. Chosen from: {\code{TRUE}, \code{FALSE}}
\item \code{include_plot_title_in_header}. Chosen from:  {\code{TRUE}, \code{FALSE}}
\item \code{always_show_fractional_seconds}. Chosen from:  {\code{TRUE}, \code{FALSE}}
\item \code{separate_date_time}. Chosen from:  {\code{TRUE}, \code{FALSE}}
\item \code{no_quotes_or_commas}. Chosen from:  {\code{TRUE}, \code{FALSE}}
\item \code{include_logger_serial}. Chosen from:  {\code{TRUE}, \code{FALSE}}
\item \code{include_sensor_serial}. Chosen from:  {\code{TRUE}, \code{FALSE}}
\item \code{date_format}. Chosen from: \code{"MDY", "YMD", "DMY"}
\item \code{date_separator}. Chosen from: {}
\item \code{time_format_24hr}. Chosen from:  {\code{TRUE}, \code{FALSE}}
\item \code{positive_number_format}. Chosen from: \code{1,2,3,4}
\item \code{negative_number_format}. Chosen from: \code{1,2,3} (corresponding to )
\item \code{include_plot_details}. Chosen from:  {\code{TRUE}, \code{FALSE}}
}
}
