% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG2.R
\name{read_fg2}
\alias{read_fg2}
\title{Read a FG2 file}
\usage{
read_fg2(filename)
}
\arguments{
\item{filename}{character, path to the file}
}
\value{
dataframe, a dataframe
}
\description{
Read data from a Geoprecision FG2 file
}
