% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geoprecision.R
\name{read_geoprecision}
\alias{read_geoprecision}
\title{Read Geoprecision file}
\usage{
read_geoprecision(filename)
}
\arguments{
\item{filename}{path to geoprecision file}
}
\description{
Read geoprecision file (autodetect GP5W / FG2)
}
