% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GP5W.R
\name{read_gp5w}
\alias{read_gp5w}
\title{Read a GP5W file}
\usage{
read_gp5w(filename)
}
\arguments{
\item{filename}{character, path to the file}
}
\value{
dataframe, a dataframe
}
\description{
Read data from a Geoprecision GP5W file
}
