% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOBO.R
\name{read_hoboware}
\alias{read_hoboware}
\title{Read Onset HOBOWare file}
\usage{
read_hoboware(filename, configuration, simplify = TRUE, encoding = "UTF-8")
}
\arguments{
\item{filename}{character, path to the file}

\item{configuration}{A named list of file configuration properties according to \code{\link{detect_hoboware_configuration}}.
If missing, LoggerReadR will attempt to auto-detect configuration properties.}

\item{simplify}{logical, whether LoggerReadR should simplify the returned dataframe (see details)}

\item{encoding}{The encoding of the textfile. Defaults to "UTF-8"}
}
\value{
dataframe, a dataframe
}
\description{
Read Onset HOBOWare file no matter how it is formatted
}
\details{
If simplify is `TRUE`, the dataframe will be standardized somewhat: row numbers and non-data columns
will be removed; time and date columns will be combined into a single column called `TIME`.
}
\author{
Nick Brown
}
