% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOBO.R
\name{read_hoboware_details}
\alias{read_hoboware_details}
\title{Read hoboware details}
\usage{
read_hoboware_details(filename, configuration, encoding = "UTF-8")
}
\arguments{
\item{filename}{character, path to the file}

\item{configuration}{A named list of file configuration properties according to \code{\link{detect_hoboware_configuration}}.
If missing, LoggerReadR will attempt to auto-detect configuration properties.}

\item{encoding}{The encoding of the textfile. Defaults to "UTF-8"}
}
\description{
Read details from a Hoboware export file if they were included
}
