% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOBO.R
\name{read_hoboware}
\alias{read_hoboware}
\title{Read Onset HOBOWare file}
\usage{
read_hoboware(filename, configuration, simplify = FALSE)
}
\arguments{
\item{filename}{character, path to the file}

\item{configuration}{A named list of file configuration properties according to \code{\link{detect_hoboware_configuration}}.
If missing, LoggerReadR will attempt to auto-detect configuration properties.}

\item{simplify}{logical, whether LoggerReadR should simplify the returned dataframe (see details)}
}
\value{
dataframe, a dataframe
}
\description{
Read Onset HOBOWare file. Optionally provide file configuration properties.
}
\details{

}
\author{
Nick Brown
}
